+++ src/main/java/com/nix/repository/CommentRepository.java
@@ -10,31 +10,31 @@ import org.springframework.data.repository.query.Param;
 
 import com.nix.models.Comment;
 
-public interface CommentRepository extends JpaRepository<Comment, Integer> {
-	public List<Comment> findByUserId(Integer userId);
+public interface CommentRepository extends JpaRepository<Comment, Long> {
+	public List<Comment> findByUserId(Long userId);
 
-	public List<Comment> findByBookId(Integer bookId);
+	public List<Comment> findByBookId(Long bookId);
 
-	public Page<Comment> findByBookId(Integer bookId, Pageable pageable);
+	public Page<Comment> findByBookId(Long bookId, Pageable pageable);
 
-	public List<Comment> findByChapterId(Integer chapterId);
+	public List<Comment> findByChapterId(Long chapterId);
 
-	public Page<Comment> findByChapterId(Integer chapterId, Pageable pageable);
+	public Page<Comment> findByChapterId(Long chapterId, Pageable pageable);
 
 	@Query("select c from Comment c where c.parentComment IS NULL")
 	public List<Comment> findParentComments();
 
 	@Query("SELECT c FROM Comment c WHERE c.parentComment IS NULL AND c.book.id = :bookId AND c.chapter.id IS NULL")
-	public Page<Comment> findParentCommentsByBookId(@Param("bookId") Integer bookId, Pageable pageable);
+	public Page<Comment> findParentCommentsByBookId(@Param("bookId") Long bookId, Pageable pageable);
 
 	@Query("SELECT c FROM Comment c WHERE c.parentComment IS NULL AND c.chapter.id = :chapterId")
-	public Page<Comment> findParentCommentsByChapterId(@Param("chapterId") Integer chapterId, Pageable pageable);
+	public Page<Comment> findParentCommentsByChapterId(@Param("chapterId") Long chapterId, Pageable pageable);
 
 	@Query("SELECT c FROM Comment c WHERE c.parentComment IS NULL AND c.post.id = :postId")
-	public Page<Comment> findParentCommentsByPostId(@Param("postId") Integer postId, Pageable pageable);
+	public Page<Comment> findParentCommentsByPostId(@Param("postId") Long postId, Pageable pageable);
 
-	public Page<Comment> findByUserIdOrderByCreatedAtDesc(Integer userId, Pageable pageable);
+	public Page<Comment> findByUserIdOrderByCreatedAtDesc(Long userId, Pageable pageable);
 
 	@Query("SELECT COUNT(c) FROM Comment c WHERE c.book.id = :bookId")
-	public Long countCommentsByBookId(@Param("bookId") Integer bookId);
+	public Long countCommentsByBookId(@Param("bookId") Long bookId);
 }
